Renamed and fixed some issues
